Refactor, make new client common sub-package
