+++ x/distribution/client/common/pretty_params.go
@@ -1,4 +1,4 @@
-package cli
+package common
 
 import (
 	"encoding/json"
